requirements.txt
# Core Dependencies
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Streamlit Dashboard
streamlit==1.29.0

# Email Support (built-in smtplib, no extra packages needed)

# Utilities
python-dateutil==2.8.2

# Optional: For advanced features
# openai==1.54.0  # If using AI features in future
